Move match-candidate partitioning code into its own module
